Add support for ESP32-C3 "contracted" with cc1101
This adds support for the Luatos ESP32C3-CORE Development Board, and
clones, wired with a CC1101 rf module.

The esp32c3 is using the native USB interface for serial console and
firmware updates, so should run on "contracted" boards without an extra
usb2serial chip.

The GDO0 and GDO2 pins are configured at IO11 and IO06 of the mcu.

Example execution command:
    pio run -e esp32c3_cdc_cc1101 -t upload -t monitor

Signed-off-by: Henk Vergonet <henk.vergonet@gmail.com>